Cut cruft & new HTTPs redirection method
In this commit, the module Web.App.Password was removed
because it was deemed not to be general enough, providing
specific functionality outside serving an HTTP app.

Additionally, the previous way to redirect HTTP traffic proved
a little problematic, so it was replaced with a simpler means:
a WAI middleware. Spawning another process would add a higher
degree of separation between secure HTTPS-land and insecure
HTTP-land. However, in testing, it would lead to zombie processes
and, if a maintainer isn't careful enough, issues with forking
and database libraries such as libPQ.