RWS-like instances, pushFunc wrapper, improved paths
Wrote MonadState & MonadWriter instances for improved generality
of RouteT. MonadState acts on the WebAppState, whereas MonadWriter
mappends Streams together to be written to the response body.

In order to decrease the out-of-control number of parameters to
runRouteT, PushFunc is in the process of being wrapped with routes
and lifted into the RouteT m s a monad transformer. This allows for
less code that is more readable.

Finally, path routing has been improved. Rather than using a couple
functions, a whole data structure has been defined. Paths are stored
in a terminal representation (i.e. as path components or as a compiled
regex) to save time when matching. This change also increases the
readability of code.