Improved CLI & scaffolding
The CLI parser was simplified. First, the actual code was refactored
for brevity and point-freedom. Second, the pid-file option and
daemonize flag were combined into one single option called daemonize.
This makes sense because it forces the user to adhere to good daemonizing
practices.

The next major improvement was the elimination of status/stop features
in the CLI. Quite frankly, these are more effectively done with a shell.
Not having to have code to do this makes the CLI significantly thinner. It
also allowed me to remove a module with less-than-pretty code. Everybody wins!